+++ common/src/main/java/xyz/eclipseisoffline/modifyplayerdata/mixin/EntityMixin.java
@@ -1,6 +1,6 @@
 package xyz.eclipseisoffline.modifyplayerdata.mixin;
 
-import net.minecraft.entity.Entity;
+import net.minecraft.world.entity.Entity;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -11,16 +11,16 @@ public abstract class EntityMixin implements VisualFireable {
 
     @Shadow
     @Final
-    protected static int ON_FIRE_FLAG_INDEX;
+    protected static int FLAG_ONFIRE;
     @Shadow
     private boolean hasVisualFire;
 
     @Shadow
-    protected abstract void setFlag(int index, boolean value);
+    protected abstract void setSharedFlag(int index, boolean value);
 
     @Override
     public void modifyPlayerData$setHasVisualFire(boolean hasVisualFire) {
         this.hasVisualFire = hasVisualFire;
-        setFlag(ON_FIRE_FLAG_INDEX, hasVisualFire);
+        setSharedFlag(FLAG_ONFIRE, hasVisualFire);
     }
 }