+++ common/src/main/java/xyz/eclipseisoffline/modifyplayerdata/mixin/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package xyz.eclipseisoffline.modifyplayerdata.mixin;
+
+import org.jspecify.annotations.NullMarked;