+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Update to Minecraft 26.1-snapshot-10.
 - Add support for the `active_effects` tag. This allows you to show an effect's icon without showing its particles, or vice versa, or to add a hidden effect to an effect instance.
+- Add support for the `recipeBook` tag.
 
 This version was not released for NeoForge, as NeoForge does not yet support this version of Minecraft.